+++ ZMacBlazor/Client/ZMachine/Instructions/MethodDescriptor.cs
@@ -5,7 +5,7 @@ namespace ZMacBlazor.Client.ZMachine.Instructions
 {
     public class MethodDescriptor
     {
-        public MethodDescriptor(ushort address, Machine machine)
+        public MethodDescriptor(int address, Machine machine)
         {
             if(machine == null) { throw new ArgumentNullException(nameof(machine)); }
 
@@ -19,22 +19,22 @@ namespace ZMacBlazor.Client.ZMachine.Instructions
             
             if(machine.Version < 5)
             {
-                var values = new ushort[LocalsCount];
+                var values = new int[LocalsCount];
                 for (var i = 0; i < LocalsCount; i++)
                 {
                     var offset = 1 + (i * 2);
                     values[i] = Bits.MakeWord(bytes.Slice(offset, 2));
                 }
                 InitialValues = values;
-                HeaderSize = (ushort)(1 + (LocalsCount * 2));
+                HeaderSize = (1 + (LocalsCount * 2));
             }
-            StartAddress = (ushort)(address + HeaderSize);
+            StartAddress = address + HeaderSize;
         }
 
-        public byte LocalsCount { get; }
-        public ushort StartAddress { get; }
-        public ICollection<ushort> InitialValues { get; } = Empty;
-        public ushort HeaderSize { get; set; } = 1;
-        readonly static ICollection<ushort> Empty = new List<ushort>();
+        public int LocalsCount { get; }
+        public int StartAddress { get; }
+        public ICollection<int> InitialValues { get; } = Empty;
+        public int HeaderSize { get; set; } = 1;
+        readonly static ICollection<int> Empty = new List<int>();
     }
 }
